backend/requirements.txt
alembic==1.16.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
Deprecated==1.2.18
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.2
